backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
chardet==3.0.4
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.0
